+++ src/main/java/racingcar/validator/NumberValidator.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.validator;
 
 import static util.CarConstant.*;
